requirements.txt
openai==1.14.3
pyttsx3==2.90
python-dotenv==1.0.1
SpeechRecognition==3.10.0
pyaudio==0.2.11
